update api request
